+++ books-and-breakfast/app/HomeScreen.tsx
@@ -36,18 +36,24 @@ function HomeScreen() {
   }, [selected]);
 
   useEffect(() => {
+    // Define an asynchronous function inside the useEffect hook to fetch the list of schools.
     const fetchSchools = async () => {
       try {
+        // Attempt to fetch the school list using the getSchoolList function.
         const schoolList = await getSchoolList();
         if (schoolList != null) {
           setSchoolOptions(schoolList);
         }
       } catch (error) {
+        // If an error occurs during fetching, log it to the console.
         console.error('Failed to fetch schools:', error);
       }
     };
+
+    // Call the fetchSchools function defined above to execute the fetching process.
+    // This function is called right after the component mounts due to the empty dependency array.
     fetchSchools();
-  }, []);
+  }, []); // The empty dependency array ensures this effect runs only once after the component mounts.
 
   // Button press handler for navigation and action buttons
   const handleButtonPress = (buttonIndex: number) => {
