+++ books-and-breakfast/app/SchoolTransportDetails.tsx
@@ -26,25 +26,52 @@ const Divider = ({ color = '#D9D9D9', thickness = 1, marginVertical = 20 }) => (
 
 export const SchoolTransportDetails = ({ schoolName }: { schoolName: string }) => {
   const [directionsInfo, setDirectionsInfo] = useState<SchoolDirections>();
+  const [isLoading, setIsLoading] = useState(true);
+  const [error, setError] = useState<string | null>(null);
 
+  // This useEffect is triggered when the component mounts and anytime the schoolName prop changes.
   useEffect(() => {
+    // Define an asynchronous function inside the effect that will fetch the school directions.
     const fetchSchoolDirections = async () => {
+      setIsLoading(true); // Set loading state to true to show a loading indicator
+      setError(null); // Reset error state to null to clear previous errors
+
       try {
+        // Attempt to fetch directions using the schoolName parameter.
         const directionsInfo = await getSchoolDirections(schoolName);
+
+        // If fetch is successful and directions are returned, update state.
         if (directionsInfo != null) {
           setDirectionsInfo(directionsInfo);
+        } else {
+          // If no data is returned, set an appropriate error message.
+          setError('No directions available for this school.');
         }
       } catch (error) {
-        console.error('Failed to fetch school directions info', error);
+        console.error('Failed to fetch school directions info', error); // Log error to console for debugging.
+        setError('Failed to load directions. Please try again later.'); // Set user-friendly error message.
+      } finally {
+        setIsLoading(false); // Ensure loading state is set to false when fetch is complete.
       }
     };
-    fetchSchoolDirections();
-  }, [schoolName]);
 
+    // Call the fetch function defined above.
+    fetchSchoolDirections();
+  }, [schoolName]); // The effect depends on schoolName, so it re-runs when schoolName changes.
+
+  // Conditionally render UI based on the state of the data fetch.
+  if (isLoading) {
+    return <Text>Loading...</Text>; // Display loading text while data is being fetched.
+  }
+  if (error) {
+    return <Text>{error}</Text>; // Show any error messages if present.
+  }
+
+  // Main content rendering, conditioned on having valid directions data.
   return (
-    <>
+    <ScrollView contentContainerStyle={styles.scrollViewContentContainer}>
       {directionsInfo && (
-        <ScrollView contentContainerStyle={styles.scrollViewContentContainer}>
+        <>
           <ArticleHeader schoolName={directionsInfo.schoolName} location={directionsInfo.address} />
           <View style={styles.section}>
             <Text style={styles.header}>Directions</Text>
@@ -64,9 +91,9 @@ export const SchoolTransportDetails = ({ schoolName }: { schoolName: string }) =
               <Text style={styles.text}>{directionsInfo.publicTransport}</Text>
             </View>
           )}
-        </ScrollView>
+        </>
       )}
-    </>
+    </ScrollView>
   );
 };
